Moved utilities file to src folder
